// src/modules/utils/enumis/topics.enum.ts
export enum SystemLogTopics {
    formSystemLogs = 'log.form.system.logs',
    getSystemLogs = 'log.get.system.logs'
}

export enum ProductAnalyticsTopics {
    fromProductsAnalytics = 'log.form.products.analytics',
    getProductsAnalytics = 'log.get.products.analytics',
    getProductWidget = 'log.get.product.widget',
    getProductsTrend = 'log.get.products.trend'
}

export enum ProfileLogTopics {
    formProfileLogs = 'log.form.profile.logs',
    getProfileLogs = 'log.get.profile.logs'
}